requirements.txt
flask
flask_mysql
Werkzeug